Solve 01-cascade-fix TOP practice
Pay attention to precedence of selectors and, if equal, pay attention to last defined selector position.
Pay attention to class vs. ID and syntax to make a child more selective.
It would be easier to write my own styles, but this was a good exercise keeping styles and modifying
selectors to fit requests.
Initially challenging, then obvious once knowledge is used.